refactor: consolidate duplicated helpers across decisions and learn pipelines
Extract three sets of duplicated code into shared locations:

1. readObservations / warnIfInvalid / writeObservations — identical in
   decisions.ts and learn.ts; export from learn.ts and import in decisions.ts.

2. _loadExistingObservations / _loadObservationsFromLog — copy-pasted between
   learning-agent.ts and decisions-agent.ts; consolidate as loadExistingObservations
   in background-runner.ts (already the shared pipeline utilities module).

3. Remove rotateLog alias from background-runner.ts — unused externally, adds
   noise without value.

Net: -90 lines of duplicated logic.
