+++ src/cli/utils/decisions-agent.ts
@@ -23,6 +23,7 @@ import { promises as fs } from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 import { promisify } from 'util';
+import { loadExistingObservations } from './background-runner.js';
 
 const execFileAsync = promisify(execFile);
 
@@ -102,7 +103,7 @@ export async function runDecisionsAgent(opts: DecisionsAgentOpts): Promise<strin
   const { cwd, dialogPairs, model, logFile, jsonHelperPath } = opts;
 
   // Load existing observations for deduplication context.
-  const existingObs = await _loadExistingObservations(jsonHelperPath, logFile, ['decision', 'pitfall']);
+  const existingObs = await loadExistingObservations(jsonHelperPath, logFile, ['decision', 'pitfall']);
 
   // Build the prompt.
   const prompt = _buildDecisionsPrompt(dialogPairs, existingObs);
@@ -315,57 +316,3 @@ function _serializePitfall(obs: RawObservation): string {
   return parts.join('; ');
 }
 
-// ---------------------------------------------------------------------------
-// Internal helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Load existing observations from the log file, filtered by types.
- * Falls back to manual log read if json-helper.cjs is unavailable.
- */
-async function _loadExistingObservations(
-  jsonHelperPath: string,
-  logFile: string,
-  types: string[],
-): Promise<string> {
-  try {
-    const { stdout } = await execFileAsync('node', [jsonHelperPath, 'filter-observations', logFile, 'confidence', '30'], {
-      timeout: 10_000,
-    });
-    const parsed: unknown = JSON.parse(stdout.trim() || '[]');
-    if (!Array.isArray(parsed)) return '[]';
-    const filtered = parsed.filter(
-      (entry): entry is Record<string, unknown> =>
-        typeof entry === 'object' && entry !== null &&
-        types.includes(String((entry as Record<string, unknown>)['type'])),
-    );
-    return JSON.stringify(filtered);
-  } catch {
-    return _loadObservationsFromLog(logFile, types);
-  }
-}
-
-async function _loadObservationsFromLog(logFile: string, types: string[]): Promise<string> {
-  try {
-    const content = await fs.readFile(logFile, 'utf-8');
-    const lines = content.split('\n').filter(Boolean);
-    const observations: unknown[] = [];
-    for (const line of lines) {
-      try {
-        const entry: unknown = JSON.parse(line);
-        if (
-          typeof entry === 'object' && entry !== null &&
-          types.includes(String((entry as Record<string, unknown>)['type'])) &&
-          (entry as Record<string, unknown>)['status'] === 'observing'
-        ) {
-          observations.push(entry);
-        }
-      } catch {
-        // Skip malformed lines.
-      }
-    }
-    return JSON.stringify(observations);
-  } catch {
-    return '[]';
-  }
-}
